Size,Tag,Color Controller Added

+++ src/Core/OnionPronia.Application/Interfaces/Services/ISizeService.cs
@@ -6,9 +6,9 @@ namespace OnionPronia.Application.Interfaces.Services
     public interface ISizeService
     {
         Task<IReadOnlyList<GetSizeItemDto>> GetAllAsync(int page, int take);
-        Task<GetSizeDto> GetByIdAsync(int? id);
+        Task<GetSizeDto> GetByIdAsync(long? id);
         Task CreateAsync(PostSizeDto sizeDto);
-        Task UpdateAsync(int id, PutSizeDto sizeDto);
-        Task DeleteAsync(int id);
+        Task UpdateAsync(long id, PutSizeDto sizeDto);
+        Task DeleteAsync(long id);
     }
 }
